Refactor code in UsePerson and FlatMapDemo classes
In UsePerson, the IntStream library was imported for creating a linked list of squares. Also, suppressed warnings for method references were added. The split string method was refactored to split at one or more whitespaces.
In FlatMapDemo, checking if a customer has orders was changed from comparing size to 0, to calling the isEmpty() method for better readability.