Refactor StringExercises tests and update/add features in FileFilterTest and LazyExceptionMessageTest
The unnecessary parts of the code inside the unit tests in 'StringExercises' class are removed, for the cleaner test structure. The functionality in those tests wasn't testing any production code but the features of Java itself.

In 'FileFilterTest', a new test method is added to explore and show example usage of Map's forEach method.

'LazyExceptionMessageTest' is a new class created to provide example tests that demonstrate how to delay the message string creation only when an assertion fails (in traditional operations, the message is always created, whether the test passes or fails). This optimization can be suitable for conditions where creating a message itself is an expensive operation.