Improve logging and stream manipulation, add assertions to test cases
The commit introduces various changes in several files.
In StringExercises.java, logging was added to different steps of the string filtering for better traceability and debugging. The logger instance is created earlier to avoid issues with sequentional operations.
In the CompanyTest.java, assertions were added for better validation of `Optional` behavior in different test cases.
An unused import in BoxedStreams.java was removed for code cleanup and line order rearranged for better readability.
In RunnableDemo.java, a sleep function was introduced to slow down the thread execution for testing purpose.
Lastly, insignificant whitespaces changing in Company.java which improve readability.