Remove MemberDAOTests and refactor DAO and tests
Deleted MemberDAOTests file because it lacked implementation. Moved the test methods to MemberDAO and created new implementation tests in InMemoryMemberDAOTests to replace them. The methods' bodies in DAO were left empty due to the lack of database. House was refactored to become an enum to prevent possible mistakes by user-input. Updated Member to a record for simplicity, auto-generation of methods, and to maintain immutability.