Update CompanyEmployee methods and tests
In CompanyEmployee.java, added a new method company() to return the Company name, making the class more robust. Modified the doWork() method to make the output more general.

In SortGolfers.java, updated the comment lines to give a clearer understanding of the operation.

A new unit test for the company() method was created in CompanyEmployeeTest.java for better coverage.