Refactor code to use streams and test cases
The code is refactored to simplify the use of arrays and lists into streams within the LoopsSortsAndIfs class. A new CompanyEmployee class was also added as well as test cases for listDirectories, listJavaSrcFiles in FileFilterTest class, and getName in CompanyEmployeeTest class to ensure the functionality works as expected. Simplified some code for readability in UsePerson class and LazyStreams classes. Various other minor improvements were made.