Refactor and add features to several classes
Several code enhancements were implemented, including the removal of an unused Logger object in LazyStreams class, a simplification in the use of isEmpty() in FlatMapDemo, and the adjustment of lambdas to method references in UsePerson. In addition, a new CompanyEmployee class was created, and a previously commented test was uncommented and fixed in the CompanyEmployeeTest class.