Refactor CompanyEmployee to Record and Update Tests
This commit refactors the CompanyEmployee to use Java records which auto-generate some methods and ensures immutability. Doing so also prompted method name changes for the Employee interface, which then led to necessary changes in the associated tests. The "getter" methods in Employee were updated to match the property names given by the new record implementation.